Added homepage in the /app/page.js instead of making it /Homepage route, added allowed image domains
