+++ client/next.config.mjs
@@ -1,7 +1,9 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   images: {
-    domains: ["encrypted-tbn0.gstatic.com", "www.winetourscapetown.com", "images.squarespace-cdn.com", "silversea-discover.imgix.net", "a.travel-assets.com,", "cdn.britannica.com", "africantravelcanvas.com", "www.tripsavvy.com"],
+    domains: ["encrypted-tbn0.gstatic.com", "www.winetourscapetown.com", "images.squarespace-cdn.com", "silversea-discover.imgix.net", "a.travel-assets.com,", "cdn.britannica.com", "africantravelcanvas.com", "www.tripsavvy.com", "i.ytimg.com"
+      ,"www.sanparks.org", "www.isimangaliso.com", "www.drakensberg.org", "www.maropeng.co.za", "www.cango-caves.co.za", "www.hluhluwegamereserve.com", "cdn.getyourguide.com", "www.isibindi.co.za", "www.nondela.com", "sanipasstours.com", "cdn-icons-png.flaticon.com", "thumbs.dreamstime.com"
+    ],
   },
 };
